requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
openai>=1.3.0
langchain>=0.0.350
langchain-community>=0.0.14
tiktoken>=0.5.1
pydantic>=2.0.0
python-docx>=1.0.0
markdown>=3.5.0
arxiv>=2.0.0
scholarly==1.7.10
newspaper3k>=0.2.8
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
rank-bm25>=0.2.2
